refactor: reply Domain Comment/ 아래로 계층 이동
