+++ src/main/java/inu/codin/codin/domain/post/domain/comment/dto/response/CommentResponseDTO.java
@@ -2,7 +2,7 @@ package inu.codin.codin.domain.post.domain.comment.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import inu.codin.codin.domain.post.domain.comment.entity.CommentEntity;
-import inu.codin.codin.domain.post.domain.reply.entity.ReplyCommentEntity;
+import inu.codin.codin.domain.post.domain.comment.reply.entity.ReplyCommentEntity;
 import inu.codin.codin.domain.post.dto.UserDto;
 import inu.codin.codin.domain.post.dto.UserInfo;
 import inu.codin.codin.domain.report.dto.response.ReportedCommentDetailResponseDTO;
