+++ src/main/java/inu/codin/codin/domain/post/domain/comment/reply/dto/request/ReplyUpdateRequestDTO.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.post.domain.reply.dto.request;
+package inu.codin.codin.domain.post.domain.comment.reply.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
