+++ src/main/java/inu/codin/codin/domain/post/domain/comment/reply/repository/ReplyCommentRepository.java
@@ -1,7 +1,7 @@
-package inu.codin.codin.domain.post.domain.reply.repository;
+package inu.codin.codin.domain.post.domain.comment.reply.repository;
 
 
-import inu.codin.codin.domain.post.domain.reply.entity.ReplyCommentEntity;
+import inu.codin.codin.domain.post.domain.comment.reply.entity.ReplyCommentEntity;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
